src/main/java/ru/nsu/ccfit/rivanov/security/login/LoginFilter.java
package ru.nsu.ccfit.rivanov.security.login;

import org.springframework.security.authentication.AuthenticationServiceException;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.AuthenticationException;
import org.springframework.security.web.authentication.AbstractAuthenticationProcessingFilter;
import org.springframework.security.web.util.matcher.RequestMatcher;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

public class LoginFilter extends AbstractAuthenticationProcessingFilter {
    public static final String LOGIN_HEADER_NAME = "login";
    public static final String PASSWORD_HEADER_NAME = "password";

    public LoginFilter(RequestMatcher requiresAuthenticationRequestMatcher) {
        super(requiresAuthenticationRequestMatcher);
    }

    @Override
    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException, IOException, ServletException {
        if (!request.getMethod().equals("POST")) {
            throw new AuthenticationServiceException("Authentication method not supported: " + request.getMethod());
        }

        String login = request.getHeader(LOGIN_HEADER_NAME);
        String password = request.getHeader(PASSWORD_HEADER_NAME);

        if (login == null) {
            login = "";
        }

        if (password == null) {
            password = "";
        }
        login = login.trim();
        UsernamePasswordAuthenticationToken authRequest = new UsernamePasswordAuthenticationToken(login, password);
        return this.getAuthenticationManager().authenticate(authRequest);
    }
}
